% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions_1_0_11.R
\name{balance}
\alias{balance}
\title{Function to create a balance table for a specified diagnostic}
\usage{
balance(input, diagnostic, approach = "none", censoring, scope,
  times.exposure, times.covariate, exposure, history = NULL,
  weight.exposure = NULL, weight.censor = NULL, strata = NULL,
  recency = NULL, average.over = NULL, periods = NULL,
  list.distance = NULL, sort.order = "alphabetical", loop = "no",
  ignore.missing.metric = "no", metric = "SMD")
}
\arguments{
\item{input}{restructured dataframe}

\item{diagnostic}{diagnostic of interest e.g. 1, 2, or 3}

\item{approach}{adjustment method e.g. "none" or "weight" or "stratify"}

\item{censoring}{use censoring indicators/weights e.g. "yes" or "no"}

\item{scope}{report the entire trellis e.g. "all", the diagonal e.g. "recent", or a summary e.g. "average"}

\item{times.exposure}{vector of exposure measurement times e.g. c(0,1,2)}

\item{times.covariate}{vector of covariate measurement times e.g. c(0,1,2)}

\item{exposure}{root name of exposure e.g. "a"}

\item{history}{root name of exposure history e.g. "h"}

\item{weight.exposure}{root name of IP exposure weights e.g. "wa"}

\item{weight.censor}{root name of IP censoring weights e.g. "ws"}

\item{strata}{root name of propensity-score strata e.g. "e"}

\item{recency}{an integer for the relative distance between exposures and covariate measurements to focus on (e.g. 0 would represent the same timing). The default is 0 for Diagnostics 1 and 3, and 1 for Diagnostic 2}

\item{average.over}{summary level for average metrics e.g. standardize over "values" or "history" or "time" or "distance"}

\item{periods}{a list of contiguous segments of relative distance to pool over e.g. list(0,1:4,5:10) would yield summaries over three segments}

\item{list.distance}{a vector of distances to retain after averaging over time e.g. c(0,2)}

\item{sort.order}{vector of root names for all covariates listed in the order in whcihc they should appear in the table (and also plot) e.g. c("n","m","o","l","p"). To display covariates in alphabetical order (the default), leave blank or type "alphabetical"}

\item{loop}{a housekeeping argument the user can ignore. It is automatically set when the balance function is called by the diagnose() function described later. The default is set to "no".}

\item{ignore.missing.metric}{"yes" or "no" depending on whether the user wishes to estimate averages over person-time when there are missing values of the mean difference or standardized mean difference. Missing values for the standardized mean difference can occur when, for example, there is no covariate variation within levels of exposure-history and measurement times. If this argument is set to "no" and there are missing values, the average will also be missing. If set to "yes" an average will be produced that ignores missing values.}
}
\description{
Function to create a balance table for a specified diagnostic
}
\examples{
balance(input, diagnostic, approach, censoring, scope, times.exposure, times.covariate, exposure, history, weight.exposure, weight.censor, strata, recency, average.over, periods, list.distance, ignore.missing.metric, loop)
}
