% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie.R
\name{cfdr_cie}
\alias{cfdr_cie}
\title{Title}
\usage{
cfdr_cie(.data, .fmla, .model = lm, ..., exposure = NULL,
  force_include = NULL, flow = "full", exponentiate = FALSE)
}
\arguments{
\item{.data}{data}

\item{.fmla}{formula}

\item{.model}{model}

\item{...}{args}

\item{exposure}{e}

\item{force_include}{g}

\item{flow}{gg}

\item{exponentiate}{gg}
}
\description{
Title
}
