% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confoundr-package.R
\docType{package}
\name{confoundr-package}
\alias{confoundr}
\alias{confoundr-package}
\title{confoundr: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
