% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions_1_0_12.R
\name{lengthen}
\alias{lengthen}
\title{Function to create a "tidy" dataframe where the key observation is the paring of exposure and covariate measurement times}
\usage{
lengthen(input, diagnostic, censoring, id, times.exposure, times.covariate,
  exposure, temporal.covariate, static.covariate = NULL, history = NULL,
  weight.exposure = NULL, censor = NULL, weight.censor = NULL,
  strata = NULL)
}
\arguments{
\item{input}{dataframe in wide format}

\item{diagnostic}{diagnostic of interest e.g. 1, 2, or 3}

\item{censoring}{use censoring indicators/weights e.g. "yes" or "no"}

\item{id}{unique observation identifier e.g. "id"}

\item{times.exposure}{a vector of exposure measurement times e.g. c(0,1,2)}

\item{times.covariate}{a vector of covariate measurement times e.g. c(0,1,2)}

\item{exposure}{the root name for exposure measurements e.g. "a"}

\item{temporal.covariate}{a vector of root names for covariates whose values change over time e.g. c("l","m","n","o","p")}

\item{static.covariate}{a vector of root names for covariates whose values do not change (covariates listed here should not appear in the temporal.covariate argument)}

\item{history}{the root name for history measurements e.g. "h"}

\item{weight.exposure}{the root name for exposure weights e.g. "wa"}

\item{censor}{the root name for censoring indicators e.g. "s"}

\item{weight.censor}{the root name for censoring weights e.g. "ws"}

\item{strata}{the root name for propensity-score strata e.g. "e"}
}
\description{
Function to create a "tidy" dataframe where the key observation is the paring of exposure and covariate measurement times
}
