% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehistory.one.R
\name{makehistory.one}
\alias{makehistory.one}
\title{Create exposure history for a single time varying exposure}
\usage{
makehistory.one(input, id, times, group = NULL, exposure,
  name.history = "h")
}
\arguments{
\item{input}{this is a matrix of}

\item{id}{vector of ids}

\item{times}{vector of times}

\item{group}{grouping variable (optional)}

\item{name.history}{}
}
\value{
A \code{\link{data.frame}} of the output stuff
}
\description{
Create exposure history for a single time varying exposure
}
\examples{
x = 5
makehistory.one(x)
\dontrun{

}
}
