% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions_1_0_12.R
\name{makeplot}
\alias{makeplot}
\title{Function to create balance plot for a specified diagnostic}
\usage{
makeplot(input, diagnostic, approach, metric = "SMD", scope,
  average.over = NULL, stratum = NULL, label.exposure = "A",
  label.covariate = "C", lbound = -1, ubound = 1, ratio = 2,
  text.axis.title = 8, text.axis.y = 6.5, text.axis.x = 6.5,
  text.strip.y = 10, text.strip.x = 10, point.size = 0.75,
  zeroline.size = 0.1, refline.size = 0.1, refline.limit.a = -0.25,
  refline.limit.b = 0.25, panel.spacing.size = 0.75, axis.title = NULL,
  label.width = 15, groupvar = "none", shape = NULL, colour = NULL,
  legend.title = "", legend.position = "bottom", text.legend = NULL)
}
\arguments{
\item{input}{output from balance()}

\item{diagnostic}{diagnostic of interest e.g. 1, 2, or 3}

\item{approach}{adjustment method e.g. "none" or "weight" or "stratify"}

\item{metric}{scale e.g. "D" for mean difference, "SMD" for standardized mean difference}

\item{scope}{report the entire trellis e.g. "all", the diagonal e.g. "recent", or a summary e.g. "average"}

\item{average.over}{level of summary for average e.g. "values" or "history" or "time" or "distance"}

\item{stratum}{the propensity-score stratum to plot}

\item{label.exposure}{common label used for exposure axis in plot (default = "A")}

\item{label.covariate}{common label used for covariate axis in plot (default = "C")}

\item{lbound}{lower bound for mean difference or standardized mean difference (default = -1)}

\item{ubound}{upper bound for mean difference or standardized mean difference (default = 1)}

\item{ratio}{aspect ratio of plot (default = 2)}

\item{text.axis.title}{font size of axis title (default = 8)}

\item{text.axis.y}{font size of y-axis values (default = 6.5)}

\item{text.axis.x}{font size of x-axis values (default = 6.5)}

\item{text.strip.y}{font size of y-axis label (default = 10)}

\item{text.strip.x}{font size of x-axis label (default = 10)}

\item{point.size}{size of data points (default = 0.75)}

\item{zeroline.size}{width of the line plotted at mean difference = 0 or standardized mean difference = 0 (default = 0.1)}

\item{refline.size}{width of the lines plotted at the specified fraction of the mean difference or standardized mean difference (default = 0.1)}

\item{refline.limit.a}{position of the lower reference line, specified as a fraction of the mean difference or standardized mean difference (default = -0.25)}

\item{refline.limit.b}{position of the upper reference line, specified as a fraction of the mean difference or standardized mean difference (default = 0.25)}

\item{panel.spacing.size}{space between each panel in the plot (default = 0.75)}

\item{axis.title}{main title for plot (optional)}

\item{label.width}{width of labels in plot (default = 15)}

\item{groupvar}{the type of grouping variable "shape" or "colour"}

\item{shape}{the variable name to assign a shape scale}

\item{colour}{the variable name to assign a color scale}

\item{legend.title}{title for legend (optional)}

\item{legend.position}{position of legend (default = "bottom")}

\item{text.legend}{text to include in legend (optional)}

\item{censoring}{use censoring indicators/weights e.g. "yes" or "no"}
}
\description{
Function to create balance plot for a specified diagnostic
}
\examples{
myplot <- makeplot(input=mytable, diagnostic=3, approach="weight", metric="SMD", scope="all", average.over, stratum, label.exposure, label.covariate, lbound, ubound, ratio, text.axis.title, text.axis.y, text.axis.x, text.strip.y, text.strip.x, point.size, zeroline.size, refline.size, refline.limit.a, refline.limit.b, panel.spacing.size, axis.title, label.width, groupvar, shape, colour, legend.title, legend.position, text.legend)
}
