% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions_1_0_12.R
\name{makehistory.one}
\alias{makehistory.one}
\title{Function to create exposure history for a single time varying exposure}
\usage{
makehistory.one(input, id, times, group = NULL, exposure,
  name.history = "h")
}
\arguments{
\item{input}{dataset in wide format}

\item{id}{unique observation identifier e.g. "id"}

\item{times}{a vector of measurement times e.g. c(0,1,2)}

\item{group}{an optional baseline variable upon which to aggregate the exposure history. This argument provides a way to adjust the metrics for a baseline covariate. For example, in the context of a trial, the grouping variable could be treatment assignment. In the context of a cohort study, this could be site e.g. "v".}

\item{exposure}{the root name for exposure e.g. "a"}

\item{name.history}{desired root name for time-indexed history variables e.g. "h"}
}
\description{
Function to create exposure history for a single time varying exposure
}
\examples{
mydata.history <- makehistory.one(input=mydata,
                                  id=id,
                                  times=c(0,1,2),
                                  exposure="a",
                                  name.history="h",
                                  group="v")
}
