% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions_1_0_12.R
\name{makehistory.two}
\alias{makehistory.two}
\title{Function to create joint exposure history for two distinct time-varying exposures}
\usage{
makehistory.two(input, id, group = NULL, exposure.a, exposure.b,
  name.history.a = "ha", name.history.b = "hb", times)
}
\arguments{
\item{input}{dataset in wide format}

\item{id}{unique observation identifier e.g. "id"}

\item{group}{an optional baseline variable upon which to aggregate the exposure history. This argument provides a way to adjust the metrics for a baseline covariate. For example, in the context of a trial, the grouping variable coul be treatment assignment. In the context of a cohort study, this could be site e.g. "v".}

\item{exposure.a}{the root name for the first exposure e.g. "a"}

\item{exposure.b}{the root name for the second exposure e.g. "z"}

\item{name.history.a}{desired root name for the first time-indexed history variables e.g. "ha"}

\item{name.history.b}{desired root name for the second time-indexed history variables e.g. "hb"}

\item{times}{a vector of measurement times e.g. c(0,1,2)}
}
\description{
Function to create joint exposure history for two distinct time-varying exposures
}
\examples{
mydata.history <- makehistory.two(input=mydata, id=id, times=c(0,1,2), exposure.a="a", exposure.b="z", name.history.a="ha", name.history.b="hb", group="v")
}
