% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions_1_0_12.R
\name{makehistory.two}
\alias{makehistory.two}
\title{Function to create joint exposure history for two distinct time-varying exposures}
\usage{
makehistory.two(input, id, group = NULL, exposure.a, exposure.b,
  name.history.a = "ha", name.history.b = "hb", times)
}
\arguments{
\item{input}{dataset in wide format}

\item{id}{unique observation identifier e.g. "id"}

\item{group}{an optional baseline variable upon which to aggregate the exposure history. This argument provides a way to adjust the metrics for a baseline covariate. For example, in the context of a trial, the grouping variable coul be treatment assignment. In the context of a cohort study, this could be site e.g. "v".}

\item{exposure.a}{the root name for the first exposure e.g. "a"}

\item{exposure.b}{the root name for the second exposure e.g. "z"}

\item{name.history.a}{desired root name for the first time-indexed history variables e.g. "ha"}

\item{name.history.b}{desired root name for the second time-indexed history variables e.g. "hb"}

\item{times}{a vector of measurement times e.g. c(0,1,2)}
}
\description{
Function to create joint exposure history for two distinct time-varying exposures
}
\examples{
# Simulate wide data set for two subjects
id <- as.numeric(c(1, 2))
a_0 <- as.numeric(c(0, 1))
a_1 <- as.numeric(c(1, 1))
a_2 <- as.numeric(c(1, 0))
z_0 <- as.numeric(c(1, 0))
z_1 <- as.numeric(c(0, 0))
z_2 <- as.numeric(c(0, 1))
l_0 <- as.numeric(rbinom(2, 1, 0.5))
l_1 <- as.numeric(rbinom(2, 1, 0.5))
l_2 <- as.numeric(rbinom(2, 1, 0.5))
m_0 <- as.numeric(rbinom(2, 1, 0.5))
m_1 <- as.numeric(rbinom(2, 1, 0.5))
m_2 <- as.numeric(rbinom(2, 1, 0.5))
n_0 <- as.numeric(rbinom(2, 1, 0.5))
n_1 <- as.numeric(rbinom(2, 1, 0.5))
n_2 <- as.numeric(rbinom(2, 1, 0.5))

mydata.wide <- data.frame(id, a_0, a_1, a_2,
                          z_0, z_1, z_2,
                          l_0, l_1, l_2,
                          m_0, m_1, m_2,
                          n_0, n_1, n_2)
# Run the makehistory.two() function
mydata.history <- makehistory.two(input=mydata.wide,
                                  id="id",
                                  times=c(0,1,2),
                                  exposure.a="a",
                                  exposure.b="z",
                                  name.history.a="ha",
                                  name.history.b="hb"
                                 )
}
