% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions_1_0_12.R
\name{omit.history}
\alias{omit.history}
\title{Function to remove irrelevant covariate history from balance tables and plots}
\usage{
omit.history(input, omission, covariate.name, distance = NULL, times = NULL)
}
\arguments{
\item{input}{restructured dataframe from lengthen()}

\item{omission}{type of omission e.g. "fixed" or "relative" or "same.time"}

\item{covariate.name}{root name of the covariate e.g. "m"}

\item{distance}{the distance between exposure and covariate measurements e.g. 2}

\item{times}{a vector of measurement times for the covariate e.g. c(1,2,3)}
}
\description{
Function to remove irrelevant covariate history from balance tables and plots
}
\examples{
mydata.long.omit <- omit.history(input=mydata.long,
                                 omission="relative",
                                 covariate.name=c("l","m","o"),
                                 distance=1)
}
